removing long typedef, all tests should pass now

+++ clang/test/CheckedCRewriter/arrtestsBUG.c
@@ -6,7 +6,7 @@
 Various array bounds tests.
 */
 
-typedef unsigned long size_t;
+typedef unsigned int size_t;
 void *memset(void * dest : byte_count(n),
              int c,
              size_t n) : bounds(dest, (_Array_ptr<char>)dest + n);
